test(media): share image runtime test fixtures
Consolidate three image runtime mock lifecycles while retaining all 42 scenarios and provider/profile-specific defaults. Reset owned mock implementations per case. Net test/support reduction: 377 lines. Related: #139428.
